chore(rimap-imap): expose DovecotHarness via test-support feature (#136)
PR12's daemon-Dovecot integration suite lives in rimap-server's tests
directory, but the harness body (compose up/down, fingerprint, ports)
is in rimap-imap. Add a `test-support` feature on rimap-imap that
re-exports `DovecotHarness` and its lifecycle so rimap-server tests
can consume it through [dev-dependencies] without duplicating the file.

The original `container.rs` mixed two concerns: container lifecycle
(stdlib + rimap-core only) and IMAP-connection helpers (rimap-audit,
rimap-config, tempfile, serde_json). Split them so only the
dependency-light half is exposed via the feature; the connection
helpers move to a sibling `connect.rs` that stays test-only.

`extern crate self as rimap_imap` aliases the current crate so
`container.rs`'s `rimap_imap::*` paths resolve when the file is
compiled as part of the lib (via `#[path]`) and when it's compiled
as a test binary (where `rimap_imap` is a real extern crate). The
optional `dep:serde_json` is pulled in by the feature so
`prune_stale_projects`'s JSON parsing keeps working without forcing
production builds to take the dep.

Refs #136.
