+++ crates/rimap-imap/src/lib.rs
@@ -17,3 +17,14 @@ pub mod types;
 pub use crate::connection::{Connection, ConnectionConfig, ImapEncryption};
 pub use crate::error::{AuthFailure, ImapError, StarttlsFailure, StarttlsRefusal};
 pub use special_use::{SpecialUse, SpecialUseMap, classify_special_use};
+
+// `test_support` re-exports `tests/integration/support/container.rs` via
+// `#[path]` so the file is the single source of truth for both in-crate
+// integration tests and the cross-crate `rimap-server` daemon-Dovecot
+// suite. The included file references `rimap_imap::*` paths as if it
+// were an external consumer; alias the current crate so those paths
+// resolve when the file is compiled as part of this lib.
+#[cfg(any(test, feature = "test-support"))]
+extern crate self as rimap_imap;
+#[cfg(any(test, feature = "test-support"))]
+pub mod test_support;
