+++ crates/rimap-imap/tests/integration/support/mod.rs
@@ -1,4 +1,5 @@
 //! Test support: Dovecot container harness (docker or podman), fixture loaders.
 
+pub mod connect;
 pub mod container;
 pub mod fixtures;
