+++ crates/rimap-server/Cargo.toml
@@ -65,6 +65,10 @@ predicates = { workspace = true }
 mail-parser = { workspace = true }
 rimap-audit = { path = "../rimap-audit", version = "1.0.0", features = ["test-injection"] }
 rimap-config = { path = "../rimap-config", version = "1.0.0", features = ["test-support"] }
+# Cross-crate test-support: rimap-imap exposes `DovecotHarness` via this
+# feature so the daemon-Dovecot integration suite can spin up a real
+# IMAP backend without copying the harness implementation.
+rimap-imap = { path = "../rimap-imap", version = "1.0.0", features = ["test-support"] }
 # Self dev-dep enables the `test-support` feature unconditionally for
 # integration tests, so `ImapMcpServer::execute_tool_for_test` is in
 # scope without every test harness passing `--features test-support`.
